backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-filter==23.1
djoser==2.1.0
Pillow==9.3.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
PyYAML==6.0
